Move ByteBuddy agent attachment into its separate class
The Hazelcast JUnit runner defines a new anonymous inner class which
depends on ByteBuddy. This is only used for during compatibility test.

However the anonymous class is loaded eagerly when the runner is loaded.
By moving the inner class into its own Utils class means it's not loaded
unless running compatibility test and I can use the runner even when
ByteBuddy is not on a classpath.

This test was failing when ByteBuddy was not on a classpath:
```java
@Test
public void foo() throws ClassNotFoundException {
  Class.forName("com.hazelcast.test.HazelcastSerialClassRunner");
}
```
